moved img processing to utils. Moved graph generation to Deck class.
