+++ ygo_small_world/config.py
@@ -1,0 +1,5 @@
+"""Small World configuration settings."""
+from collections import namedtuple
+
+Settings = namedtuple("Settings", ["card_size", "max_pixel_brightness"])
+SETTINGS = Settings(624, 255)
